Deserialized JSON
Added classes to extract data from JSON, as well as print statements to show everything that was retrieved in JSON format as something that is human readable. This may need changed in the future if the professor determines that he wants the link from the assignment example, but it shouldn't be too hard.